refactor: address clippy warnings and simplify docs
- Use SystemParam derive in gen plugin to avoid too_many_arguments
- Replace ALL_EDITABLE_FILES constant with is_editable_file() helper
- Fix disk space check overflow by using u128 for statvfs math
- Simplify GEMINI.md and add AGENTS.md to reference CLAUDE.md
- Add -D warnings flag to clippy command in docs

+++ crates/cli/src/cli/doctor.rs
@@ -653,8 +653,9 @@ fn check_disk_space() -> CheckResult {
             unsafe {
                 let mut stat: nix::libc::statvfs = std::mem::zeroed();
                 if nix::libc::statvfs(path.as_ptr(), &mut stat) == 0 {
-                    let free_bytes = stat.f_bavail * stat.f_frsize;
-                    let free_mb = free_bytes / (1024 * 1024);
+                    // statvfs field widths vary by platform; normalize before math.
+                    let free_bytes = (stat.f_bavail as u128) * (stat.f_frsize as u128);
+                    let free_mb = free_bytes / ((1024_u128) * (1024_u128));
 
                     if free_mb < 100 {
                         return CheckResult::warn(
